backend/requirements.txt
python-pptx
pdfplumber
python-docx